requirements.txt
numpy~=1.26
pandas~=2.1
scikit-learn~=1.3
joblib~=1.3
pyyaml~=6.0
websockets~=11.0
httpx~=0.27
pydantic~=2.6
matplotlib~=3.8
aiofiles~=23.2
